app/backend-django/requirements.txt
django>=5.1,<5.2
Pillow>=10,<12
dj-database-url>=2,<3
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
channels[daphne]>=4.1,<5
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"
